refactor: extract TraceContextBase for tracing contexts
