+++ nautilus/src/nautilus/tracing/ExceptionBasedTraceContext.hpp
@@ -1,22 +1,18 @@
-
 #pragma once
 
 #include "ExecutionTrace.hpp"
+#include "TraceContextBase.hpp"
 #include "TraceOperation.hpp"
 #include "nautilus/CompilableFunction.hpp"
 #include "nautilus/common/FunctionAttributes.hpp"
 #include "nautilus/options.hpp"
 #include "nautilus/tracing/TracingInterface.hpp"
-#include "tag/Tag.hpp"
 #include "tag/TagRecorder.hpp"
 #include <cstddef>
 #include <cstdint>
-#include <cstring>
 #include <functional>
 #include <list>
 #include <memory>
-#include <optional>
-#include <unordered_map>
 #include <unordered_set>
 #include <vector>
 
@@ -28,175 +24,6 @@ namespace nautilus::tracing {
 class ExecutionTrace;
 class SymbolicExecutionContext;
 class TraceModule;
-struct StaticVarHolder {
-	explicit StaticVarHolder(const void* ptr, size_t size) : ptr(ptr), size(size) {
-	}
-
-private:
-	const void* ptr;
-	size_t size;
-	friend uint64_t hashStaticVector(const std::vector<StaticVarHolder>& data);
-	friend size_t getStaticVarValue(const StaticVarHolder& holder);
-};
-
-inline size_t getStaticVarValue(const StaticVarHolder& holder) {
-	size_t result = 0;
-	std::memcpy(&result, holder.ptr, holder.size);
-	return result;
-}
-
-/**
- * @brief Efficiently tracks reference counts and computes an incremental hash of alive variables.
- *
- * This class maintains reference counts using a sparse hash map to support the full uint32_t range
- * of variable IDs while keeping memory usage reasonable. The hash reflects both which variables
- * are alive and their reference counts, updated incrementally in O(1) average time.
- *
- * Implementation details:
- * - Uses XOR-based hashing for O(1) incremental updates
- * - Each variable ID is mixed with a constant multiplier for better hash distribution
- * - The hash incorporates both variable identity (ID) and reference count
- * - Uses unordered_map for sparse storage (only allocates for variables that exist)
- * - Entries are erased when their count reaches zero, so the map stays bounded by the
- *   peak number of alive variables instead of growing with every value ref ever seen
- *
- * Performance characteristics:
- * - increment(): O(1) average - hash map lookup + two XOR operations, two multiplications
- * - decrement(): O(1) average - hash map lookup + two XOR operations, two multiplications
- * - hash(): O(1) - returns cached value
- *
- * @note Changed from fixed array to hash map to support uint32_t ValueRef (was uint16_t).
- */
-class AliveVariableHash {
-	static constexpr uint64_t HASH_MULTIPLIER = 0x9e3779b97f4a7c15; // Golden ratio constant for good mixing
-
-	std::unordered_map<uint32_t, uint32_t> counts;
-	uint64_t alive_hash = 0;
-
-public:
-	/**
-	 * @brief Default constructor. No initialization needed as counts are zero-initialized.
-	 */
-	AliveVariableHash() = default;
-
-	/**
-	 * @brief Increments the reference count for a variable and updates the hash.
-	 *
-	 * The hash is updated by XOR-ing out the old contribution ((id * HASH_MULTIPLIER) * old_count)
-	 * and XOR-ing in the new contribution ((id * HASH_MULTIPLIER) * new_count).
-	 *
-	 * @param id Variable identifier (32-bit value)
-	 */
-	inline void increment(uint32_t id) noexcept {
-		uint32_t& c = counts[id];
-		alive_hash ^= (id * HASH_MULTIPLIER) * c;
-		++c;
-		alive_hash ^= (id * HASH_MULTIPLIER) * c;
-	}
-
-	/**
-	 * @brief Decrements the reference count for a variable and updates the hash.
-	 *
-	 * The hash is updated by XOR-ing out the old contribution ((id * HASH_MULTIPLIER) * old_count)
-	 * and XOR-ing in the new contribution ((id * HASH_MULTIPLIER) * new_count).
-	 *
-	 * Entries that reach a count of zero are erased. This is hash-neutral (a zero count
-	 * contributes 0 to the XOR hash, identical to an absent entry) and keeps the map size
-	 * proportional to the number of currently-alive variables. Without the erase, the map
-	 * grows with every value ref ever seen and — because the trace context is thread_local —
-	 * persists across trace iterations and across traced functions.
-	 *
-	 * @param id Variable identifier (32-bit value)
-	 */
-	inline void decrement(uint32_t id) noexcept {
-		const auto it = counts.try_emplace(id, 0).first;
-		uint32_t& c = it->second;
-		alive_hash ^= (id * HASH_MULTIPLIER) * c;
-		--c;
-		alive_hash ^= (id * HASH_MULTIPLIER) * c;
-		if (c == 0) {
-			counts.erase(it);
-		}
-	}
-
-	/**
-	 * @brief Returns the current hash value representing the state of alive variables.
-	 *
-	 * The hash reflects both which variables have non-zero reference counts and the
-	 * magnitude of those counts. This value is maintained incrementally and can be
-	 * retrieved in O(1) time.
-	 *
-	 * @return 64-bit hash value representing current variable state
-	 */
-	inline uint64_t hash() const noexcept {
-		return alive_hash;
-	}
-
-	/**
-	 * @brief Returns the number of tracked entries.
-	 *
-	 * Since zero-count entries are erased by decrement(), this is the number of
-	 * currently-alive variables (variables with a non-zero reference count).
-	 *
-	 * @return Number of entries in the underlying map
-	 */
-	inline size_t size() const noexcept {
-		return counts.size();
-	}
-
-	/**
-	 * @brief Resets all reference counts and hash to initial state.
-	 *
-	 * This efficiently clears all counts without creating a temporary object.
-	 * Since decrement() erases entries when they reach zero, a balanced trace iteration
-	 * leaves the map empty and this is a no-op. The emptiness check (rather than a hash
-	 * check) also guards against the rare XOR collision where live entries hash to 0.
-	 */
-	inline void reset() noexcept {
-		if (!counts.empty()) {
-			counts.clear();
-		}
-		alive_hash = 0;
-	}
-};
-
-/**
- * @brief State that requires initialization for tracing operations.
- * This is initialized in the trace context when tracing begins and reset when it ends.
- * Holds references to stack-allocated objects.
- */
-struct TraceState {
-	TagRecorder& tagRecorder;
-	ExecutionTrace& executionTrace;
-	SymbolicExecutionContext& symbolicExecutionContext;
-	const engine::Options& options;
-	std::unordered_map<void*, uint32_t> normalizedFunctionNameCache; // Maps function pointers to normalized indices
-	uint32_t nextNormalizedFunctionIndex = 0;                        // Counter for normalized function names
-
-	TraceState(TagRecorder& tr, ExecutionTrace& et, SymbolicExecutionContext& sec, const engine::Options& opts);
-};
-
-/**
- * @brief Common base class for trace context implementations.
- *
- * Holds shared state (TraceState, mangledNameCache) and provides
- * getMangledName() / getFunctionName() so they are not duplicated
- * across ExceptionBasedTraceContext and LazyTraceContext.
- */
-class TraceContextBase : public TracingInterface {
-public:
-	~TraceContextBase() override = default;
-
-	std::string getMangledName(void* fnptr);
-	std::string getFunctionName(void* fnptr, const std::string& mangledName);
-
-protected:
-	// Injected state - holds references to stack-allocated objects (ExecutionTrace, SymbolicExecutionContext).
-	// Empty when not tracing and stored inline to avoid a per-trace heap allocation.
-	std::optional<TraceState> state;
-
-	std::unordered_map<void*, std::string> mangledNameCache;
-};
 
 /**
  * @brief The trace context manages a thread local instance to record a symbolic execution trace of a given Nautilus
@@ -271,9 +98,6 @@ public:
 	TypedValueRef& traceNautilusFunctionPtr(const NautilusFunctionDefinition* definition,
 	                                        std::function<void()> fwrapper) override;
 
-	void pushStaticVal(void* ptr, size_t size) override;
-	void popStaticVal() override;
-
 	// --- Non-interface public API ---
 
 	~ExceptionBasedTraceContext() override = default;
@@ -324,11 +148,7 @@ private:
 	template <typename OnCreation>
 	TypedValueRef& traceOperation(Op op, OnCreation&& onCreation);
 	Snapshot recordSnapshot();
-	std::string formatStaticVars() const;
 
-	// Persistent state - reset between trace iterations via resume()
-	std::vector<StaticVarHolder> staticVars; // Tracks static variable states for snapshot hashing
-	AliveVariableHash aliveVars;             // Tracks alive variables with incremental hash (256KB)
 	std::list<compiler::CompilableFunction> functionsToTrace = std::list<compiler::CompilableFunction> {};
 	std::unordered_set<std::string> registeredFunctions;
 };
